Add IP address of first master as a variable that Ansible can use
When we move Ansible to supporting multiple clusters, there is no
simple way to get the first IP address of the master for the current
cluster.

Instead of doing that, we can simply have terraform add the IP address
to a separate variable in the inventory file, and use that variable in
the two locations where we need the IP address.