Weave: Split into separate configuration for workers & masters.
We are going to use the master servers for forming the quorum for
allocating IP address ranges, with the worker nodes asking the masters
for IP address ranges.

This should allow us to use a relatively static set of servers for the
Weave peer list, which will become useful when we start using Weave
across multiple clusters.