Weave: Use masters to allocate IP address ranges.
This patch changes Weave to only require a quorum between the master
nodes when allocating IP addresses. When a new worker node starts up,
it will ask the master nodes to allocate an IP range for the node.

The goal of this change is to later change Weave to bootstrap using
only the list of master nodes. This list should be relatively static,
and should make it easy to spread Weave across multiple clusters.